backend/requirements.txt
anyio==4.9.0
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==7.0.1
Django==5.1.5
django-allauth==65.3.1
django-cors-headers==4.6.0
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-spectacular==0.28.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logging==0.4.9.6
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.1.0
promise==2.3
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.11.2
python-socketio==5.12.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
service-identity==24.2.0
setuptools==75.8.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
text-unidecode==1.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wsproto==1.2.0
zope.interface==7.2
